iox-#33 moved named pipe into platforms, introduced NamedPipeSender
